+++ src/main/kotlin/com/sendy/application/usecase/transfer/TransferService.kt
@@ -3,7 +3,7 @@ package com.sendy.application.usecase.transfer
 import com.sendy.application.dto.transfer.TransferMoneyCommand
 import com.sendy.application.usecase.transfer.command.TransferMoneyUseCase
 import com.sendy.domain.enum.TransferStatusEnum
-import com.sendy.domain.transfer.TransactionHistoryRepository
+import com.sendy.domain.account.TransactionHistoryRepository
 import com.sendy.domain.transfer.Transfer
 import com.sendy.domain.transfer.TransferRepository
 import com.sendy.infrastructure.persistence.transfer.event.TransferProcessed
