+++ src/main/kotlin/com/sendy/domain/account/TransactionHistoryRepository.kt
@@ -1,4 +1,4 @@
-package com.sendy.domain.transfer
+package com.sendy.domain.account
 
 interface TransactionHistoryRepository {
     fun save(domain: TransactionHistory)
